Change data and indicator type from double to float

+++ alis/src/main/java/com/alphatica/alis/data/loader/ohlcv/OHLCVRow.java
@@ -14,24 +14,24 @@ import static com.alphatica.alis.data.layer.Layer.OPEN;
 public class OHLCVRow {
 
 	private Time time;
-	private List<Double> data;
+	private List<Float> data;
 
 	public static OHLCVRow fromParts(String date, String open, String high, String low, String close, String volume) {
 		OHLCVRow quote = new OHLCVRow();
 		quote.data = new ArrayList<>();
 		quote.time = new Time(Integer.parseInt(date.replace(" ", "").replace("-", "")));
-		quote.data.add(Double.parseDouble(open));
-		quote.data.add(Double.parseDouble(high));
-		quote.data.add(Double.parseDouble(low));
-		quote.data.add(Double.parseDouble(close));
-		quote.data.add(Double.parseDouble(volume) * (quote.get(OPEN) + quote.get(HIGH) + quote.get(LOW) + quote.get(CLOSE)) / 4.0);
-		quote.data.add(-1.0);
-		quote.data.add(-1.0);
-		quote.data.add(-1.0);
+		quote.data.add(Float.parseFloat(open));
+		quote.data.add(Float.parseFloat(high));
+		quote.data.add(Float.parseFloat(low));
+		quote.data.add(Float.parseFloat(close));
+		quote.data.add(Float.parseFloat(volume) * (quote.get(OPEN) + quote.get(HIGH) + quote.get(LOW) + quote.get(CLOSE)) / 4.0f);
+		quote.data.add(-1.0f);
+		quote.data.add(-1.0f);
+		quote.data.add(-1.0f);
 		return quote;
 	}
 
-	public double get(Layer layer) {
+	public float get(Layer layer) {
 		return data.get(layer.getIndex());
 	}
 
@@ -40,7 +40,7 @@ public class OHLCVRow {
 	}
 
 	public void parseAndSet(String text, Layer layer) {
-		this.data.set(layer.getIndex(), Double.parseDouble(text));
+		this.data.set(layer.getIndex(), Float.parseFloat(text));
 	}
 
 }
